wishlist/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.4.0
contourpy==1.2.1
cycler==0.12.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.12.1
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.4
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1.post1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==2.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.10
protobuf==4.25.5
Pygments==2.18.0
Pympler==1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pytest==8.3.5
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scipy==1.14.0
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.3
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
drf-yasg==1.21.7

